requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
pandas>=2.2.2
tzdata>=2024.1   # ensures IST timezone works on Windows
